Refactors type imports and reorganizes type definitions
Reorganizes type definitions by moving and renaming type files for better modularity and clarity. Updates type imports across the codebase to reflect the new structure, ensuring consistency and maintainability.

Groups related types into dedicated files (e.g., `tokens`, `figma`, `results`, `astro`, `ui`) and removes redundant or misplaced definitions. Improves code readability and simplifies future type management.
