+++ src/lint/colors/test-paint-style.ts
@@ -1,8 +1,9 @@
 import { addResult } from "../results";
 import { astroColorIsUsingCorrectTheme } from "./tests/astro-color-is-using-correct-theme";
-import { AstroComponent } from "../components/types";
-import { AstroTheme, LintingResult } from "../types";
-import { FillStyleNode } from "../../types";
+import { AstroComponent } from "../../types/astro";
+import { AstroTheme } from "../../types/tokens";
+import { FillStyleNode } from "../../types/figma";
+import { LintingResult } from "../../types/results";
 import { usingAstroColor } from "./tests/using-astro-color";
 import { usingColorFromComponent } from "./tests/using-color-from-component";
